task 8/task 8/requirements.txt
streamlit
groq
pinecone-client
sentence-transformers
python-dotenv
